requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.3
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.14.0
gunicorn==22.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.1
